+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @regenrek/gpuix-react
 
+## 0.5.0-regenrek.5
+
+### Patch Changes
+
+- Export the browser-surface event types and JSX intrinsic contract from the React package.
+- Updated dependencies
+  - @regenrek/gpuix-native@0.5.0-regenrek.5
+
 ## 0.5.0-regenrek.4
 
 ### Minor Changes